refactor(storage): dashboard settings 3件を SettingsRepository に置換（PBI 01 Phase2-4）
